+++ src/app/api/app.h
@@ -96,8 +96,8 @@ public:
     //轨迹配置
     static int XWidth;         //水平
     static int YHeight;        //纵向
-    static float ZeroPointX;         //零点X
-    static float ZeroPointY;        //零点Y
+    static double ZeroPointX;         //零点X
+    static double ZeroPointY;        //零点Y
 
     static void readConfig();       //读取配置文件,在main函数最开始加载程序载入
     static void writeConfig();      //写入配置文件,在更改配置文件程序关闭时调用
